pyproject.toml
[tool.poetry]
name = "hw1"
version = "0.1.0"
description = ""
authors = ["Haohang Li <hli113@stevens.edu>"]
readme = "README.md"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[tool.poetry.dependencies]
python = ">=3.10, < 3.11"
numpy = "^1.26.0"
pillow = "^10.0.1"
typer = "^0.9.0"
rich = "^13.6.0"
black = "^23.9.1"
ipykernel = "^6.25.2"
opencv-python = "^4.8.1.78"
numba = "^0.58.1"
scipy = "^1.11.3"
tqdm = "^4.66.1"
matplotlib = "^3.8.0"
scikit-learn = "^1.3.2"
jax = {extras = ["cpu"], version = "^0.4.20"}


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
